+++ package.json
@@ -3,6 +3,10 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@craco/craco": "^6.2.0",
+        "@testing-library/jest-dom": "^5.11.4",
+        "@testing-library/react": "^11.1.0",
+        "@testing-library/user-event": "^12.1.10",
         "@types/classnames": "^2.2.11",
         "@types/deep-equal": "^1.0.1",
         "@types/jest": "^26.0.15",
@@ -17,16 +21,13 @@
         "antd": "^4.16.10",
         "antd-theme-webpack-plugin": "^1.3.9",
         "axios": "^0.21.1",
-        "babel-plugin-import": "^1.13.3",
         "classnames": "^2.2.6",
-        "customize-cra": "^1.0.0",
+        "craco-antd": "^1.19.0",
         "deep-equal": "^2.0.5",
         "moment": "^2.29.1",
         "prettier": "^2.2.1",
         "query-string": "^6.14.0",
         "react": "^17.0.2",
-        "react-app-rewire-less": "^2.1.3",
-        "react-app-rewired": "^2.1.8",
         "react-dom": "^17.0.2",
         "react-markdown": "^6.0.3",
         "react-redux": "^7.2.2",
@@ -37,6 +38,7 @@
         "remark-gfm": "^1.0.0",
         "tslint": "^6.1.3",
         "typescript": "^4.1.5",
+        "web-vitals": "^1.0.1",
         "yaml": "^1.10.0"
     },
     "scripts": {
@@ -44,20 +46,29 @@
         "formatter-write": "prettier --write './src/**/*.ts*'",
         "tslint": "tslint -c tslint.json -p tsconfig.json",
         "tslint-fix": "tslint --fix -c tslint.json -p tsconfig.json",
-        "start": "react-app-rewired start",
-        "build": "react-app-rewired build",
-        "test": "react-app-rewired test",
+        "start": "craco start",
+        "build": "craco build",
+        "test": "craco test",
         "eject": "react-scripts eject"
     },
     "eslintConfig": {
-        "extends": "react-app"
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
     },
-    "browserslist": [
-        ">0.2%",
-        "not dead",
-        "not ie <= 11",
-        "not op_mini all"
-    ],
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
     "devDependencies": {
         "tslint-config-prettier": "^1.18.0"
     }
